Fixes logging
