+++ coinbase-integration/adapter/src/main/resources/logback.xml
@@ -2,11 +2,9 @@
 
     <!-- APPENDERS -->
     <appender name="STDOUT" class="ch.qos.logback.core.ConsoleAppender">
-        <layout class="ch.qos.logback.classic.PatternLayout">
-            <Pattern>
-                %cyan(%d{HH:mm:ss.SSS}) %gray([%thread]) %highlight(%-5level) %magenta(%logger{36}) - %msg%n
-            </Pattern>
-        </layout>
+        <encoder>
+            <pattern>%cyan(%d{HH:mm:ss.SSS}) %white([%thread]) %highlight(%-5level) %magenta(%logger{36}) - %msg%n</pattern>
+        </encoder>
     </appender>
 
     <!-- LOGGERS -->
