+++ coinbase-integration/build.gradle
@@ -36,7 +36,6 @@ subprojects {
 
         // Slf4JLogging
         implementation "org.slf4j:slf4j-api:$slf4jVersion"
-        implementation "org.slf4j:slf4j-simple:$slf4jVersion"
 
         // RethinkDB
         implementation "com.rethinkdb:rethinkdb-driver:$rethinkdbVersion"
