+++ coinbase-integration/subscriber/src/main/kotlin/me/pysquad/cryptobot/subscriber/models.kt
@@ -1,8 +1,6 @@
 package me.pysquad.cryptobot.subscriber
 
 import com.fasterxml.jackson.annotation.JsonInclude
-import com.fasterxml.jackson.databind.PropertyNamingStrategies
-import com.fasterxml.jackson.databind.annotation.JsonNaming
 import org.http4k.core.Body
 import org.http4k.format.Jackson.auto
 
@@ -22,25 +20,4 @@ data class SubscriberResponse(
         fun failure(type: String, message: String, reason: String) =
             SubscriberResponse(type = type, message = message, reason = reason, data = emptyList())
     }
-}
-
-@JsonNaming(PropertyNamingStrategies.SnakeCaseStrategy::class)
-data class CoinbaseWsSubscribeRequest(
-    val type: String,
-    val productIds: ProductIds,
-    val channels: List<String>
-)
-
-@JsonNaming(PropertyNamingStrategies.SnakeCaseStrategy::class)
-data class CoinbaseWsResponse(
-    var type: String = "",
-    var message: String = "",
-    var reason: String = "",
-    var channels: List<CoinbaseWsChannel> = emptyList()
-) {
-    @JsonNaming(PropertyNamingStrategies.SnakeCaseStrategy::class)
-    data class CoinbaseWsChannel(
-        var name: String = "",
-        var productIds: ProductIds = emptyList()
-    )
 }
