+++ coinbase-integration/subscriber/src/main/resources/logback.xml
@@ -1,23 +1,15 @@
 <configuration>
 
     <!-- APPENDERS -->
-    <appender name="ASYNC" class="ch.qos.logback.classic.AsyncAppender">
-        <appender name="STDOUT" class="ch.qos.logback.core.ConsoleAppender">
-            <encoder class="ch.qos.logback.classic.encoder.PatternLayoutEncoder">
-                <Pattern>
-                    %cyan(%d{HH:mm:ss.SSS}) %white([%thread]) %highlight(%-5level) %magenta(%logger{36}) - %msg%n
-                </Pattern>
-            </encoder>
-        </appender>
+    <appender name="STDOUT" class="ch.qos.logback.core.ConsoleAppender">
+        <encoder>
+            <pattern>%cyan(%d{HH:mm:ss.SSS}) %white([%thread]) %highlight(%-5level) %magenta(%logger{36}) - %msg%n</pattern>
+        </encoder>
     </appender>
 
     <!-- LOGGERS -->
-    <logger name="me.pysquad.cryptobot.subscriber" additivity="false" level="DEBUG">
-        <appender-ref ref="ASYNC"/>
-    </logger>
-
-    <logger name="me.pysquad.cryptobot.http4k.starter" additivity="false" level="DEBUG">
-        <appender-ref ref="ASYNC"/>
+    <logger name="me.pysquad.cryptobot" additivity="false" level="DEBUG">
+        <appender-ref ref="STDOUT"/>
     </logger>
 
 </configuration>
