+++ http4k-starter/build.gradle
@@ -28,7 +28,6 @@ dependencies {
 
     // Slf4JLogging
     implementation "org.slf4j:slf4j-api:$slf4jVersion"
-    implementation "org.slf4j:slf4j-simple:$slf4jVersion"
 }
 
 publishing {
