+++ http4k-starter/src/main/kotlin/me/pysquad/cryptobot/http4k/starter/Common.kt
@@ -5,12 +5,13 @@ import org.http4k.contract.bindContract
 import org.http4k.contract.contract
 import org.http4k.core.*
 import org.http4k.filter.RequestFilters
+import org.http4k.filter.ServerFilters
+import org.http4k.format.Jackson.auto
 import org.http4k.server.Http4kServer
 import org.http4k.server.Jetty
 import org.http4k.server.asServer
 import org.slf4j.Logger
 import org.slf4j.LoggerFactory
-import org.http4k.format.Jackson.auto
 
 typealias ContractRoutes = MutableList<ContractRoute>
